.src/requirements.txt
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
idna==2.7
pyobjc==4.2.2
pyobjc-core==4.2.2
pyobjc-framework-Accounts==4.2.2
pyobjc-framework-AddressBook==4.2.2
pyobjc-framework-AppleScriptKit==4.2.2
pyobjc-framework-AppleScriptObjC==4.2.2
pyobjc-framework-ApplicationServices==4.2.2
pyobjc-framework-Automator==4.2.2
pyobjc-framework-AVFoundation==4.2.2
pyobjc-framework-AVKit==4.2.2
pyobjc-framework-BusinessChat==4.2.2
pyobjc-framework-CalendarStore==4.2.2
pyobjc-framework-CFNetwork==4.2.2
pyobjc-framework-CloudKit==4.2.2
pyobjc-framework-Cocoa==4.2.2
pyobjc-framework-Collaboration==4.2.2
pyobjc-framework-ColorSync==4.2.2
pyobjc-framework-Contacts==4.2.2
pyobjc-framework-ContactsUI==4.2.2
pyobjc-framework-CoreBluetooth==4.2.2
pyobjc-framework-CoreData==4.2.2
pyobjc-framework-CoreLocation==4.2.2
pyobjc-framework-CoreML==4.2.2
pyobjc-framework-CoreServices==4.2.2
pyobjc-framework-CoreSpotlight==4.2.2
pyobjc-framework-CoreText==4.2.2
pyobjc-framework-CoreWLAN==4.2.2
pyobjc-framework-CryptoTokenKit==4.2.2
pyobjc-framework-DictionaryServices==4.2.2
pyobjc-framework-DiskArbitration==4.2.2
pyobjc-framework-EventKit==4.2.2
pyobjc-framework-ExceptionHandling==4.2.2
pyobjc-framework-ExternalAccessory==4.2.2
pyobjc-framework-FinderSync==4.2.2
pyobjc-framework-FSEvents==4.2.2
pyobjc-framework-GameCenter==4.2.2
pyobjc-framework-GameController==4.2.2
pyobjc-framework-GameKit==4.2.2
pyobjc-framework-GameplayKit==4.2.2
pyobjc-framework-ImageCaptureCore==4.2.2
pyobjc-framework-IMServicePlugIn==4.2.2
pyobjc-framework-InputMethodKit==4.2.2
pyobjc-framework-InstallerPlugins==4.2.2
pyobjc-framework-InstantMessage==4.2.2
pyobjc-framework-Intents==4.2.2
pyobjc-framework-IOSurface==4.2.2
pyobjc-framework-iTunesLibrary==4.2.2
pyobjc-framework-LatentSemanticMapping==4.2.2
pyobjc-framework-LaunchServices==4.2.2
pyobjc-framework-libdispatch==4.2.2
pyobjc-framework-LocalAuthentication==4.2.2
pyobjc-framework-MapKit==4.2.2
pyobjc-framework-MediaAccessibility==4.2.2
pyobjc-framework-MediaLibrary==4.2.2
pyobjc-framework-MediaPlayer==4.2.2
pyobjc-framework-ModelIO==4.2.2
pyobjc-framework-MultipeerConnectivity==4.2.2
pyobjc-framework-NetFS==4.2.2
pyobjc-framework-NetworkExtension==4.2.2
pyobjc-framework-NotificationCenter==4.2.2
pyobjc-framework-OpenDirectory==4.2.2
pyobjc-framework-Photos==4.2.2
pyobjc-framework-PhotosUI==4.2.2
pyobjc-framework-PreferencePanes==4.2.2
pyobjc-framework-PubSub==4.2.2
pyobjc-framework-QTKit==4.2.2
pyobjc-framework-Quartz==4.2.2
pyobjc-framework-SafariServices==4.2.2
pyobjc-framework-SceneKit==4.2.2
pyobjc-framework-ScreenSaver==4.2.2
pyobjc-framework-ScriptingBridge==4.2.2
pyobjc-framework-SearchKit==4.2.2
pyobjc-framework-Security==4.2.2
pyobjc-framework-SecurityFoundation==4.2.2
pyobjc-framework-SecurityInterface==4.2.2
pyobjc-framework-ServiceManagement==4.2.2
pyobjc-framework-Social==4.2.2
pyobjc-framework-SpriteKit==4.2.2
pyobjc-framework-StoreKit==4.2.2
pyobjc-framework-SyncServices==4.2.2
pyobjc-framework-SystemConfiguration==4.2.2
pyobjc-framework-Vision==4.2.2
pyobjc-framework-WebKit==4.2.2
requests==2.19.1
urllib3==1.23
